Separate "about this documentation" from user tasks
Make the first topic more of an "about" topic
to position OpenJ9, what it is, how we present
information about Java versions, what we support
and where to find more information.

Move the task-oriented sections to a new topic that
we can build out as more content is written to support
configuration and tuning.

Also fix a link error in the version0.10.md topic.

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>