Correct 0.12.0 release line to say yes for 8 & 11
It had incorrectly said "yes" only for JDK12 which won't ship until
March.  Update the "No" for JDK8 and JDK11 to say "Yes" based on the
statements by RedHat that they will step forward as the stewards of
JDK8 and by others to indicate they will help with backports.

I've left the `(*2)` note on both of those "Yes"s though.

Updated the other "No(*2)"'s as well

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>