Add draft release notes
To help users understand which features for 0.12.0
are already in the OpenJ9 code, we are introducing
draft release notes. This will be useful for users
who have picked up the last release from Adopt
(what is not in their build), but also useful for
users who pick up our planned milestone builds.

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>