Finalise 0.12.0 release topics
Updates to the what's new ahead of the 0.12.0
release to remove references to the release being
in progress. Update the support page to include
the macOS build system for 8 and remove the note
from 0.12.0 release to say it is in plan.

Late change to include RSA algorithm support added.

Pointed the 11 and 12 release topics at the
Adopt archive for macOS builds.

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>