Update for new staging service and draft site
A new staging service is now available that can
host multiple pull requests. Update the
contribution guidelines with the new repository
and URL for previewing drafts.

Draft doc is now hosted at the ghpages branch
of the repo, with the last release published
at the Eclipse site. Update the README to reflect
this.

Partially completes work for #176

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>