Updated Diag overview to support new structure
The Diagnostics section is being expanded to cover all
data, tools, and interfaces supported by OpenJ9. This
requires a re-org of the current structure and an update
to the overview topic (renamed from Tools and data).

This is part of a larger piece of work, broken down into
incremental steps.

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>