Support for terabytes in -X and -XX options
This fix adds support for 'T' and 't' suffix (indicating terabytes)
to -X and -XX options that take  a <size> parameter, for OpenJ9
0.18.0 and above versions.

Removed mention of this being only supported on 64-bit, after getting
feedbacks from reviewers on the code commits.

Fixed the sample in 'Using -XX command-line options' which was a -X
based sample and replaced it with an -XX example.

Also updated the 0.18.0 what's new and its table of contents.

Fixes: #429
[skip ci]

Signed-off-by: Alexandre Vermeerbergen <avermeerbergen@gmail.com>