Support for terabytes in -X and -XX options
This fix adds support for 'T' and 't' suffix (indicating terabytes)
to -X and -XX options that take  a <size> parameter, for OpenJ9
0.18.0 and above versions.

Also mentions that this added support only applies to 64-bit versions
of OpenJ9.

Fixed the sample in 'Using -XX command-line options' which was a -X
based sample and replaced it with an -XX example.

Includes update of what's new in 0.18.0

Fixes: #429
[skip ci]