Correct documentation regarding XcompilationThreads option
Existing documentation specifies that the number of compilation threads
needs to be in the 1-4 interval. Starting with Java8 we have increased the
maximum number of compilation threads to 7.
Also, the behavior has changed in that the JVM will no longer fail if the
user specifies an invalid number of compilation threads; instead it will use
the default number according to its internal heuristics.

Closes #435

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>