Apply CSS changes for updated dependency levels
OpenJ9-docs build tools were updated earlier this week to move
up dependency levels. A new docker image was created and pushed
to the Docker hub Eclipse area. Draft docs are already built with
these levels, but required CSS changes.

The openj9-docs-0.17.0 branch was then updated to add in Google
Analytics tagging, but the HTML was rebuilt with the later build tool
levels and without the required CSS changes, causing the docs to render
incorrectly.

The same CSS changes are now applied to the branch. This has been
tested with a local build successfully.

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>