Document jdmpview native library location options
Document the jdmpview options that are used to find native libraries
when looking at a core file on a different system from where it was
produced, or when the JVM that produced the core is no longer at the
original location.

Also remove references to XML files, which are obsolete and no longer
produced by jextract.

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>