Remove sentence about default behaviour of shared classes
https://github.com/eclipse/openj9-docs/issues/569
Removed sentence about default behaviour of shared classes as there's already a link to the "Class data sharing" topic, which describes that. Instead, expanded on the info the user can get from that topic. This way, the user can still easily get the shared classes info but the -Xaot topic won't need to be updated (possibly getting missed again and providing incorrect info) when/if the default shared classes behaviour changes in the future.

Signed-off-by: Esther Dovey <doveye@uk.ibm.com>