Add initial "Release note" topic for 0.28
Add the initial draft "Release note" topic for 0.28, ready for updates for each feature as it comes.
This release is to support JDK 17 (not LTS), so add content to mention that.

Also move one of the previous "Release note" topics into the "Earlier releases" section to avoid cluttering the table of contents.

Signed-off-by: Esther Dovey <doveye@uk.ibm.com>