Update supported platforms topic for 0.35.0
https://github.com/eclipse-openj9/openj9-docs/issues/1008

Updated the "Eclipse OpenJ9 releases" table. Removed the note (1) saying that 0.35 is a future release. Added more future releases and removed older ones.
Removed the "OpenJDK 18" sections.
Added RHEL9 to the list of supported platforms for Linux
Incorporated review feedback
Added a note for RHEL 9.0 and Ubuntu 22.04 distributions on AArch64

Signed-off-by: Sreekala Gopakumar <sreekala.gopakumar@ibm.com>