Update requirements.txt
Jenkins build was failing because of the following error:
pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
poetry 1.8.1 requires packaging>=23.1, but you have packaging 22.0 which is incompatible.
Updated the Requirements.txt with packaging==23.2

Signed-off-by: Sreekala Gopakumar <sreekala.gopakumar@ibm.com>