Update z/OS MAXSHARE recommendations for persistent shared cache
The MAXSHARE option controls the amount of storage that can be used for
mmap64. This parameter applies on a per address space level. We can
reduce the current recommendation of 1GB to 300MB to better align with
the SCC default usage requirements. Also, add text to discuss JOBNAME
filter and suppression of IEF043I diagnostic messsages from appearing.

Signed-off-by: Joran Siu <joransiu@ca.ibm.com>